Identified Issue: When encoding/decoding numeric values that have a decimal portion of 0 (eg 1.0, 2.0, 123422.0), the encoder will strip off the trailing 0 and will write the value as an Integer.  So when decoding using the array/dictionary helpers, the value will be read as an Integer.
Added: Added support to exclude top level properties when doing dictionary decoding.  This is helpful when you may be decoding and object that has known and unknown properties.  You can use the standard method for decoding the known properties and then use the dictionary decode to load only the unknown.